refactor: reuse code editor
